requirements.txt
requests
kafka-python
pyspark
textblob
psycopg2-binary
streamlit
pandas
plotly
tensorflow